Transition from classic debian packaging to stdeb
When attempting to publish the source package to a launchpad PPA
the debian packaging was not configured properly for building the
package via automatic PPA builds.

This transition replaces the debian directory with a stdeb.cfg file
which accomplishes much the same as was configured in the debian
directory.  This stdeb packaging method was already codified in the
Makefile and it was able to build a locally installable .deb, but was
not configured properly for ubuntu PPA building.